+++ pdf/lib/src/graphics.dart
@@ -199,6 +199,12 @@ class PdfGraphics {
     buf.putString("$x $y m\n");
   }
 
+  /// Draw a bézier curve
+  void curveTo(
+      double x1, double y1, double x2, double y2, double x3, double y3) {
+    buf.putString("$x1 $y1 $x2 $y2 $x3 $y3 c\n");
+  }
+
   void drawShape(String d) {
     var sb = StringBuffer();